+++ fuzz/fuzz_targets/aur_version.rs
@@ -1,5 +1,7 @@
 #![no_main]
 
+use libfuzzer_sys::fuzz_target;
+
 use blueline::version::{AurVersionInfo, VersionInfo};
 
 fuzz_target!(|data: &[u8]| {